move Login component into Auth
